refactor(spright): extract chat auto-scroll into AutoScrollManager
